+++ libft.h
@@ -108,7 +108,7 @@ t_list			*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
 char			*ft_itoa_base(long long value, int base);
 char			*ft_strjoin_d(char **s1, char **s2, int del);
 int				get_next_line(int fd, char **line);
-int				ft_cntchr(char *str, char c);
+ssize_t			ft_cntchr(char *str, char c);
 char			*ft_realloc(char **str, size_t new_len);
 void			**ft_newarr(size_t num);
 size_t			ft_arrlen(void **arr);
